[MIG] base_location: Don't need cities for migrating
Previous script required to have cities populated for working, but that's not
usual thing, as people in v11 may choose to not import them, and for older databases
they even weren't that option.

With this improve script, now we populate city table for ZIP entries without city,
so the rest of the queries are properly executed in any case.